Eliminate some memmove/memcpy
This restores the multiply and division dualnumber operation
sequence to what it was before #11, however, with the key check
of whether the two operands are the same object. If they are,
it is a bug to first multiply(divide) the derivatives by the value
and then execute another statement. In this latter case, you really
do need a "temporary" to hold the expression.

In the former case, however, creating a "temporary" is unnecessarily
expensive, especially in the case of static derivative storage.
For static derivative storage, it doesn't matter if you have a
function with move semantics...you're still going to be copying
its contents in the body of operator+(DN &&, const DN &)
in the case of multiplication or operator-(DN &&,const DN &) in
the case of division.