Add boolean template parameter to DualNumber
This template parameter allows the user to specify at compile
time whether they want to allow skipping of derivative computation.
If they do not want to allow it, then we write the boolean checks
in such a way that the compiler will be able to optimize it away,
leaving previous performance untouched.