Unit test options for disabling long double tests
`METAPHYSICL_RUN='libtool --mode=execute valgrind' make check` fails on
systems with 80-bit long double, because valgrind doesn't support 80-bit
arithmetic and forces a bunch of calculations down to 64-bit instead.
But with this commit, we can `METAPHYSICL_OPTIONS=--disable-long-double`
too and at least get valgrind to run on the float and double (and of
course various integer) tests.