When we exceed our container size, we'll throw
The big question here was on whether this change would incur too much
expense ... and the answer appears to be no. In a run of our virtual
pebble-bed FHR, pre-throw took 37.1 seconds and post-throw took 37.7
seconds, so probably even within the noise. Moreover, the profile didn't
show `resize` as being a function of import; the flat percent
contribution was 0.053%.