Initialize values in our test instantiations
We do want to allow our numeric classes to have uninitialized
instantiations, because every operation here counts and because we want
them to behave like built-in numeric types in that respect.

But in a test where we're going to *output* their values, we really need
to instantiate them.

This fixes a bug for me where every several dozen runs was kicking up
SIGFPE.